+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/address/response/ResGetDefaultAddressDto.java
@@ -1,0 +1,17 @@
+package com.sparta.hotdeal.user.application.dtos.address.response;
+
+import java.util.UUID;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ResGetDefaultAddressDto {
+    private UUID addressId;
+    private String zipNum;
+    private String city;
+    private String district;
+    private String streetName;
+    private String streetNum;
+    private String detailAddr;
+}
