+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/address/response/ResPatchDefaultAddressByIdDto.java
@@ -1,0 +1,11 @@
+package com.sparta.hotdeal.user.application.dtos.address.response;
+
+import java.util.UUID;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ResPatchDefaultAddressByIdDto {
+    private UUID addressId;
+}
