+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/auth/request/ReqPostLoginDto.java
@@ -1,0 +1,11 @@
+package com.sparta.hotdeal.user.application.dtos.auth.request;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ReqPostLoginDto {
+    private String email;
+    private String password;
+}
