+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/auth/request/ReqPostSignUpDto.java
@@ -1,0 +1,12 @@
+package com.sparta.hotdeal.user.application.dtos.auth.request;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ReqPostSignUpDto {
+    private String email;
+    private String password;
+    private String nickname;
+}
