+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/auth/request/ReqPostVerifyEmailDto.java
@@ -1,0 +1,10 @@
+package com.sparta.hotdeal.user.application.dtos.auth.request;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ReqPostVerifyEmailDto {
+    private String email;
+}
