+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/auth/response/ResPostLoginDto.java
@@ -1,0 +1,11 @@
+package com.sparta.hotdeal.user.application.dtos.auth.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ResPostLoginDto {
+    private String accessToken;
+    private String refreshToken;
+}
