+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/users/request/ReqPatchUsersInfoByIdDto.java
@@ -1,0 +1,10 @@
+package com.sparta.hotdeal.user.application.dtos.users.request;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ReqPatchUsersInfoByIdDto {
+    private String nickname;
+}
