+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/users/request/ReqPatchUsersPasswordByIdDto.java
@@ -1,0 +1,4 @@
+package com.sparta.hotdeal.user.application.dtos.users.request;
+
+public class ReqPatchUsersPasswordByIdDto {
+}
