+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/users/response/ResGetUsersByIdDto.java
@@ -1,0 +1,14 @@
+package com.sparta.hotdeal.user.application.dtos.users.response;
+
+import java.util.UUID;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ResGetUsersByIdDto {
+    private UUID userId;
+    private String email;
+    private String nickname;
+    private UUID defaultAddressId;
+}
