+++ user/src/main/java/com/sparta/hotdeal/user/application/dtos/users/response/ResPatchUsersPasswordByIdDto.java
@@ -1,0 +1,11 @@
+package com.sparta.hotdeal.user.application.dtos.users.response;
+
+import java.util.UUID;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ResPatchUsersPasswordByIdDto {
+    private UUID userId;
+}
