--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/ResponseDto.java
@@ -1,18 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ResponseDto<T> {
-    private String message;
-    private T data;
-
-    public static <T> ResponseDto<T> of(String message, T data) {
-        return ResponseDto.<T>builder()
-                .message(message)
-                .data(data)
-                .build();
-    }
-}
