--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/auth/request/ReqPostCheckEmailDto.java
@@ -1,10 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos.auth.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ReqPostCheckEmailDto {
-    private String email;
-}
