--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/auth/request/ReqPostConfirmEmailDto.java
@@ -1,11 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos.auth.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ReqPostConfirmEmailDto {
-    private String email;
-    private String verificationToken;
-}
