--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/auth/request/ReqPostSignUpDto.java
@@ -1,12 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos.auth.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ReqPostSignUpDto {
-    private String email;
-    private String password;
-    private String nickname;
-}
