--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/auth/response/ResPostLoginDto.java
@@ -1,11 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos.auth.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ResPostLoginDto {
-    private String accessToken;
-    private String refreshToken;
-}
