--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/auth/response/ResPostSignUpDto.java
@@ -1,11 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos.auth.response;
-
-import java.util.UUID;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ResPostSignUpDto {
-    private UUID userId;
-}
