--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/users/request/ReqPatchUsersInfoByIdDto.java
@@ -1,10 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos.users.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ReqPatchUsersInfoByIdDto {
-    private String nickname;
-}
