--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/users/request/ReqPatchUsersPasswordByIdDto.java
@@ -1,4 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos.users.request;
-
-public class ReqPatchUsersPasswordByIdDto {
-}
