--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/users/response/ResDeleteUsersByIdDto.java
@@ -1,11 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos.users.response;
-
-import java.util.UUID;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ResDeleteUsersByIdDto {
-    private UUID userId;
-}
