--- user/src/main/java/com/sparta/hotdeal/user/presentation/dtos/users/response/ResGetUsersByIdDto.java
@@ -1,14 +1,0 @@
-package com.sparta.hotdeal.user.presentation.dtos.users.response;
-
-import java.util.UUID;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ResGetUsersByIdDto {
-    private UUID userId;
-    private String email;
-    private String nickname;
-    private UUID defaultAddressId;
-}
